pyproject.toml
[project]
name = "ensemble-claude"
version = "0.5.0"
description = "AI Orchestration Tool for Claude Code - Multi-agent orchestration for complex development tasks"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Chika Kakazu", email = "chika@example.com" },
]
keywords = ["ai", "orchestration", "claude", "multi-agent", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
ensemble = "ensemble.cli:main"

[project.urls]
Homepage = "https://github.com/ChikaKakazu/ensemble"
Documentation = "https://github.com/ChikaKakazu/ensemble#readme"
Repository = "https://github.com/ChikaKakazu/ensemble"
Issues = "https://github.com/ChikaKakazu/ensemble/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ensemble"]
# Include non-Python template files
artifacts = [
    "*.md",
    "*.yaml",
    "*.sh",
]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src/ensemble --cov-report=term-missing"

[tool.coverage.run]
source = ["src/ensemble"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
